experimentation-playground/pyproject.toml
[project]
name = "experimentation-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.3.5",
    "datasets>=4.4.1",
    "notebook>=7.5.0",
    "openai>=2.8.1",
    "tqdm>=4.67.1",
    "voyageai>=0.3.5",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.74.1",
]

[project.scripts]
run = "experimentation_playground.cli:cli"
